extract input layer into its own file
